@stschiff 2021-08-03
OK wow. This is all very interesting, but I'd like to understand it better. We should talk about it in person! I'll take a closer look as well.