@stschiff 2021-08-08
I've at least conceptually come up with a substantially faster algorithm which should make the coding and recoding steps unnecessary. I can't wait to try that out.