@stschiff 2021-08-13
OK, I've looked through this and decided I'll go on a separate branch. There are a number of steps involved to get this coded up, starting with some more tests and things. I would suggest we keep this PR open, as the information that the Exception-throwing takes such a massive performance footprint is definitely something we don't want to forget.