@stschiff 2021-08-13
Started to work on this now in #118