@stschiff 2021-08-18
I will close this. It was super helpful for figuring out the massive bottleneck in the recode steps. I suggest we keep the branch around.